+++ src/test/mocha-browser.d.ts
@@ -1,6 +1,6 @@
 declare module 'mocha/mocha.js' {
   const mocha: {
-    setup(options: { ui: 'tdd'; reporter?: undefined }): void;
+    setup(options: { ui: 'tdd'; reporter?: undefined; timeout?: number }): void;
     run(callback: (failures: number) => void): void;
   };
   export default mocha;